+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.37.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.37.0)
     - `csvtk cut`:
         - allow setting `-f/--fields` multiple times, useful for selecting a long column list.
+    - `csvtk csv2jon`:
+        - fix `-b/--blanks` and `-n/--parse-num`. [#352](https://github.com/shenwei356/csvtk/issues/352)
+        - add a new flag `-K/--skip-key` to skip putting KEY in the data when using -k KEY. [#351](https://github.com/shenwei356/csvtk/issues/351)
 - [csvtk v0.36.0](https://github.com/shenwei356/csvtk/releases/tag/v0.36.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.36.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.36.0)
     - new command `csvtk shuf`: shuffle rows. [#151](https://github.com/shenwei356/csvtk/issues/151)